+++ src/Symfony/Component/Validator/Tests/Constraints/ValidValidatorTest.php
@@ -20,7 +20,7 @@ class ValidValidatorTest extends TestCase
     public function testPropertyPathsArePassedToNestedContexts()
     {
         $validatorBuilder = new ValidatorBuilder();
-        $validator = $validatorBuilder->enableAnnotationMapping()->getValidator();
+        $validator = $validatorBuilder->enableAttributeMapping()->getValidator();
 
         $violations = $validator->validate(new Foo(), null, ['nested']);
 
@@ -31,7 +31,7 @@ class ValidValidatorTest extends TestCase
     public function testNullValues()
     {
         $validatorBuilder = new ValidatorBuilder();
-        $validator = $validatorBuilder->enableAnnotationMapping()->getValidator();
+        $validator = $validatorBuilder->enableAttributeMapping()->getValidator();
 
         $foo = new Foo();
         $foo->fooBar = null;